Rename logic.js to HealthEstimate.js
