+++ src/module/HealthEstimate.js
